Move base64 from tls to crypto
